refactor: [TEO-376] move download logic to own function, remove TODOs regarding getManyForDownload interfaces
