+++ frontend/src/services/access-requests/access-requests.api.ts
@@ -59,7 +59,6 @@ export const getRejectedAccessRequests = async (
   }).then((res) => res.data);
 };
 
-// TODO: turn parameters into an interface
 export const downloadAccessRequests = async (
   status: RequestStatus,
   orderBy?: string,
